requirements.txt
Django~=4.1
djangorestframework~=3.13.1
drf-yasg~=1.21.3
django-cors-headers~=3.13.0
djangorestframework-simplejwt~=5.2.0
django-colorfield~=0.7.2